requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
opencv-python==4.10.0.84
modelscope==1.20.0
pyclipper==1.3.0.post6
shapely==2.0.6
packaging==24.2
pyyaml==6.0.2
addict==2.4.0
oss2==2.19.1
datasets==2.16.0
torch==2.2.2
pillow==11.0.0
simplejson==3.19.3
sortedcontainers==2.4.0
numpy==1.24.1
python-dotenv==1.0.1
python-multipart==0.0.17


